requirements.txt
numpy
pandas
matplotlib
plotly
kaleido
polyline
streamlit
gurobipy
pyomo
googlemaps
rasterio
openrouteservice
millify
seaborn
scipy
requests
setuptools
python-dotenv